backend/pyproject.toml
[project]
name = "statement-categorizer-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlmodel>=0.0.21",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",          # async SQLite driver
    "asyncpg>=0.29.0",            # async PostgreSQL driver (for server mode)
    "greenlet>=3.0.0",            # required by SQLAlchemy async on Python 3.13
    # Config
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    # File handling
    "python-multipart>=0.0.9",    # FastAPI file uploads
    "aiofiles>=23.2.0",           # async local file I/O
    # Statement parsing
    "pdfplumber>=0.11.0",         # PDF text + table extraction
    "pandas>=2.2.0",              # CSV / tabular data processing
    "ofxparse>=0.21",             # OFX / QFX bank export parsing
    # LLM categorizer
    "anthropic>=0.36.0",          # Claude API (used when LLM_BACKEND=claude)
    "httpx>=0.27.0",              # HTTP client for Ollama API calls
    # S3 storage (optional — only used when STORAGE_BACKEND=s3)
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",              # TestClient
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]        # pycodestyle, pyflakes, isort

[tool.pytest.ini_options]
asyncio_mode = "auto"

# Dev helper scripts — run with: python -m scripts.<name>
# e.g. python -m scripts.seed_categories
